moved keyboard to beet maker
